Reorganized test packages
