+++ src/test/java/uk/ac/ebi/ampt2d/accession/AccessioningServiceGeneratorBInjectionTest.java
@@ -24,8 +24,8 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.TestPropertySource;
 import org.springframework.test.context.junit4.SpringRunner;
-import uk.ac.ebi.ampt2d.accession.generator.TestAccessionGeneratorB;
-import uk.ac.ebi.ampt2d.accession.generator.TestsGeneratorConfiguration;
+import uk.ac.ebi.ampt2d.accession.test.generator.TestAccessionGeneratorB;
+import uk.ac.ebi.ampt2d.accession.test.configuration.TestsGeneratorConfiguration;
 
 import java.util.Collections;
 import java.util.Map;
