+++ src/test/java/uk/ac/ebi/ampt2d/accession/test/generator/TestAccessionGeneratorB.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  *
  */
-package uk.ac.ebi.ampt2d.accession.generator;
+package uk.ac.ebi.ampt2d.accession.test.generator;
 
 public class TestAccessionGeneratorB extends TestPrefixAccessionGenerator {
