+++ src/test/java/uk/ac/ebi/ampt2d/accession/test/generator/TestPrefixAccessionGenerator.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  *
  */
-package uk.ac.ebi.ampt2d.accession.generator;
+package uk.ac.ebi.ampt2d.accession.test.generator;
 
 import uk.ac.ebi.ampt2d.accession.AccessionGenerator;
